refactor: update request data class
