+++ adcio_analytics/src/main/java/ai/corca/adcio_analytics/network/data/event/AnalyticsPageViewRequest.kt
@@ -1,4 +1,4 @@
-package ai.corca.adcio_analytics.network.data.pageview
+package ai.corca.adcio_analytics.network.data.event
 
 import com.google.gson.annotations.SerializedName
 
@@ -7,8 +7,8 @@ data class AnalyticsPageViewRequest(
     @SerializedName("deviceId") val deviceId: String,
     @SerializedName("customerId") val customerId: String?,
     @SerializedName("storeId") val storeId: String,
-    @SerializedName("productCode") val productCode: String?,
+    @SerializedName("productIdOnStore") val productIdOnStore: String?,
     @SerializedName("path") val path: String,
-    @SerializedName("title") val title: String,
+    @SerializedName("title") val title: String?,
     @SerializedName("referrer") val referrer: String?,
 )
